// src/services/generated/model/account-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface AccountDto {
    'id'?: string;
    'name'?: string;
    'accountType'?: AccountDtoAccountTypeEnum;
    'currency'?: AccountDtoCurrencyEnum;
    'accountContainer'?: string;
    'balance'?: number;
}

export const AccountDtoAccountTypeEnum = {
    Checking: 'CHECKING',
    Savings: 'SAVINGS',
    Cash: 'CASH'
} as const;

export type AccountDtoAccountTypeEnum = typeof AccountDtoAccountTypeEnum[keyof typeof AccountDtoAccountTypeEnum];
export const AccountDtoCurrencyEnum = {
    Chf: 'CHF',
    Eur: 'EUR',
    Huf: 'HUF',
    Usd: 'USD'
} as const;

export type AccountDtoCurrencyEnum = typeof AccountDtoCurrencyEnum[keyof typeof AccountDtoCurrencyEnum];